+++ src/sql/run/resultset.py
@@ -97,7 +97,9 @@ class ResultSet(ColumnGuesserMixin):
         """Store the DB fetched results into the internal list of results"""
         to_add = self._config.displaylimit - len(self._results)
         self._results.extend(elements)
-        self._pretty_table.add_rows(elements[:to_add])
+        self._pretty_table.add_rows(
+            elements if self._config.displaylimit == 0 else elements[:to_add]
+        )
 
     def mark_fetching_as_done(self):
         self._mark_fetching_as_done = True